requirements.txt
pytz~=2021.1
requests~=2.25.1
xmltodict~=0.12.0
fastapi==0.63.0
uvicorn==0.11.8
gunicorn
boto3==1.17.5
pygelf==0.4.0
PyYAML